requirements.txt
alabaster==0.7.12
amqp==2.6.1
appdirs==1.4.3
argon2-cffi==20.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.5.0
astroid==2.11.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
binaryornot==0.4.4
black==22.3.0
bleach==5.0.0
blinker==1.4
boto3==1.20.21
botocore==1.23.21
cairocffi==1.2.0
CairoSVG==2.5.2
cbor==1.0.0
celery==4.4.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.9
chrome-gnome-shell==0.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
crispy-bootstrap5==0.6
cryptography==3.4.7
cssselect2==0.4.1
cupshelpers==1.0
Cython==0.29.14
daphne==2.4.1
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.44.0
django-appconf==1.0.4
django-celery-beat==2.2.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-coverage-plugin==2.0.2
django-crispy-forms==1.11.1
django-debug-toolbar==3.2.4
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.2
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.3.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-modeltranslation==0.15.1
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-nine==0.2.5
django-on-heroku==1.1.2
django-redis==4.12.1
django-resized==0.3.11
django-rest-swagger==2.2.0
django-rosetta==0.9.4
django-select2==7.4.2
django-sms==0.5.0
django-storages==1.12.3
django-stubs==1.9.0
django-stubs-ext==0.4.0
django-tables2==2.3.3
django-templated-mail==1.1.1
django-tempus-dominus==5.1.2.13
django-timezone-field==4.1.2
django-weasyprint==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djoser==2.1.0
docopt==0.6.2
docutils==0.17.1
drf-base64==2.0
drf-nested-serializer==0.1.0
drf-orjson-renderer==1.6.0
drf-yasg==1.20.0
elasticsearch==7.17.3
elasticsearch-dsl==7.4.0
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.3
factory-boy==3.2.1
Faker==13.3.4
fastjsonschema==2.15.3
filelock==3.0.12
flake8==4.0.1
flake8-isort==4.1.1
flower==0.9.7
FormEncode==2.0.0
future==0.18.2
gunicorn==20.1.0
hiredis==1.1.0
html5lib==1.1
httplib2==0.14.0
humanize==3.4.1
hyperlink==19.0.0
idna==2.10
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
incremental==16.10.1
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
ipdb==0.13.9
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
jsonparser==1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyring==18.0.1
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
livereload==2.6.3
lz4==3.0.2+dfsg
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.2.0
mypy==0.942
mypy-extensions==0.4.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.11
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
openpyxl==2.4.11
orjson==3.6.8
packaging==20.9
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pbr==5.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.1.0
pipenv==11.9.0
pipreqs==0.4.11
platformdirs==2.5.2
pluggy==1.0.0
polib==1.1.1
poyo==0.5.0
prometheus-client==0.8.0
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.20
pycups==1.9.73
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.1.0
pylint==2.13.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
Pyphen==0.10.0
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.18.1
pytest==7.1.1
pytest-sugar==0.9.4
python-apt==2.0.0+ubuntu0.20.4.5
python-crontab==2.5.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-slugify==4.0.1
python-snappy==0.5.3
python3-openid==3.2.0
PyTrie==0.2
pytz==2022.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
redis==3.2.1
reportlab==3.5.34
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.5.0
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
simplejson==3.17.6
six==1.15.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.8.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
tablib==3.0.0
termcolor==1.1.0
terminado==0.13.3
testfixtures==6.18.5
testresources==2.0.1
text-unidecode==1.3
tinycss2==1.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
twilio==7.3.2
Twisted==18.9.0
txaio==2.10.0
types-pytz==2021.3.6
types-PyYAML==6.0.6
typing-extensions==4.2.0
u-msgpack-python==2.1
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.4
uritemplate==3.0.1
urllib3==1.26.4
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
watchdog==2.1.7
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==5.2.0
widgetsnbextension==3.6.0
wrapt==1.14.1
wsaccel==0.6.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yarg==0.1.9
youtube-dl==2020.3.24
zipp==3.8.0
zope.interface==4.7.1
